Fixed a bug with category loading.  Previously, a method in a category that overrode a method in the superclass would only be picked up by subclasses that were not yet initialized.
Quentin: This was the category loading problem that you saw.  It was dependent upon load order, so was difficult to reproduce, however a Smalltalk category on NSString overriding a method from NSObject demonstrated it (because GSString was always initialized by the time Smalltalk code is loaded because LKCompiler use it, but not by the time ObjC code is loaded).  



git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/libobjc2/trunk@29347 72102866-910b-0410-8b05-ffd578937521