Initial (working) type feedback pass.  Records the IMP for every message lookup.  Note: This writes 3 words for every single message send, which means that it makes code VERY slow and can fill up the filesystem if you leave it enabled.
The aim of this pass is to identify message send sites which can benefit from speculative inlining and inline caching.  By itself, this pass is useless.  It will later be combined with one that parses this information and then uses the existing caching and inlining infrastructure to optimize hot message send sites.

Now, however, it's time for me to go and dance, so the useful bit will have to wait.



git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/libobjc2/trunk@30259 72102866-910b-0410-8b05-ffd578937521