Tweak auto-copy code.  Now it only copies classes that are explicitly registered to be copied on heap assignment.  By default, this is just _NSConcreteStackBlock.  Other classes can be registered (LanguageKit should register BlockClosure to make sure that Smalltalk works).
Fred: This makes the stuff that we discussed briefly at FOSDEM possible in GC
mode: We can allocate a GSStackEvent subclass of NSEvent on the stack.  If it
implements a -copy method that returns an NSEvent and is registered with the
runtime in this way, then any code that assigns it anywhere on the heap will
end up implicitly creating a heap copy.



git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/libobjc2/trunk@33198 72102866-910b-0410-8b05-ffd578937521